# api/openapi.yaml
openapi: 3.1.0
info:
  title: Stoganet api-proxy
  version: 0.1.0
  description: Aggregating backend for Stoganet client apps. Current surface = auth + healthz.

servers:
  - url: https://api.stoganet.com

components:
  securitySchemes:
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - invalid_credentials
                - account_locked
                - token_expired
                - token_invalid
                - jellyfin_session_expired
                - backend_unavailable
                - item_not_found
                - rate_limited
                - validation_failed
                - internal
            message: { type: string }
            details: { nullable: true }
        request_id: { type: string }

    User:
      type: object
      required: [id, email, display_name]
      properties:
        id:           { type: string }
        email:        { type: string }
        display_name: { type: string }

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:     { type: string, minLength: 1 }
        password:     { type: string, minLength: 1 }
        device_label: { type: string, nullable: true }

    TokenPair:
      type: object
      required: [access_token, refresh_token, user]
      properties:
        access_token:  { type: string }
        refresh_token: { type: string }
        user:          { $ref: '#/components/schemas/User' }

    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }

    QuickConnectStartResponse:
      type: object
      required: [code, poll_token]
      properties:
        code:       { type: string }
        poll_token: { type: string }

    QuickConnectPollRequest:
      type: object
      required: [poll_token]
      properties:
        poll_token: { type: string }

    MediaType:
      type: string
      enum: [movie, tv]

    MediaState:
      type: string
      enum: [playable, downloading, requestable]

    PlayInfo:
      type: object
      required: [jellyfin_item_id, jellyfin_base_url, jellyfin_access_token, jellyfin_user_id]
      properties:
        jellyfin_item_id:      { type: string }
        jellyfin_base_url:     { type: string }
        jellyfin_access_token: { type: string }
        jellyfin_user_id:      { type: string }

    CastMember:
      type: object
      required: [name, role]
      properties:
        name: { type: string }
        role: { type: string }

    CatalogItem:
      type: object
      required: [id, title, year, type, poster, overview, state]
      properties:
        id:       { type: string }
        title:    { type: string }
        year:     { type: integer }
        type:     { $ref: '#/components/schemas/MediaType' }
        poster:   { type: string }
        backdrop: { type: string, nullable: true }
        overview: { type: string }
        state:    { $ref: '#/components/schemas/MediaState' }

    CatalogDetail:
      allOf:
        - $ref: '#/components/schemas/CatalogItem'
        - type: object
          required: [genres, runtime, cast, seasons]
          properties:
            genres:  { type: array, items: { type: string } }
            runtime: { type: integer, description: Duration in minutes }
            cast:    { type: array, items: { $ref: '#/components/schemas/CastMember' } }
            seasons: { type: integer, description: Number of seasons; 0 for movies }
            play:    { $ref: '#/components/schemas/PlayInfo', nullable: true }

    CatalogListResponse:
      type: object
      required: [items, total]
      properties:
        items:       { type: array, items: { $ref: '#/components/schemas/CatalogItem' } }
        total:       { type: integer }
        next_cursor: { type: string, nullable: true }

paths:
  /healthz:
    get:
      summary: Liveness probe
      operationId: getHealthz
      responses:
        '200':
          description: Process alive
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [ok] }

  /auth/login:
    post:
      summary: Username + password login (proxied through Jellyfin AuthenticateByName)
      operationId: postAuthLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          description: Bad credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '423':
          description: Account locked due to repeated failures
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: Jellyfin unavailable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /auth/refresh:
    post:
      summary: Refresh token rotation
      operationId: postAuthRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          description: Refresh token invalid / reused / expired
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /auth/logout:
    post:
      summary: Revoke a refresh token
      operationId: postAuthLogout
      security: [ { BearerJWT: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '204': { description: Revoked }

  /auth/logout/all:
    post:
      summary: Revoke all refresh tokens for the user
      operationId: postAuthLogoutAll
      security: [ { BearerJWT: [] } ]
      responses:
        '204': { description: Revoked }

  /auth/quick-connect/start:
    post:
      summary: Begin a Jellyfin Quick Connect handshake
      operationId: postAuthQuickConnectStart
      responses:
        '200':
          description: Code issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuickConnectStartResponse' }
        '503':
          description: Jellyfin unavailable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /auth/quick-connect/poll:
    post:
      summary: Poll Quick Connect approval
      operationId: postAuthQuickConnectPoll
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuickConnectPollRequest' }
      responses:
        '200':
          description: Approved; token pair issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '202':
          description: Still waiting
        '410':
          description: Expired or rejected
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /catalog/{id}:
    get:
      summary: Item detail with playback handoff
      operationId: getCatalogId
      security: [ { BearerJWT: [] } ]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Item detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogDetail' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Item not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: Jellyfin unavailable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /catalog:
    get:
      summary: Paginated catalog browse
      operationId: getCatalog
      security: [ { BearerJWT: [] } ]
      parameters:
        - name: type
          in: query
          required: false
          schema: { $ref: '#/components/schemas/MediaType' }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 40, minimum: 1, maximum: 100 }
        - name: cursor
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Catalog page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogListResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: Jellyfin unavailable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }